restructuring to remove bloat
